STOR-2954: move tls utils
